Android Extracted FFT Data Test
Import data from Android. Our previous commits have been building up to this: instead of processing raw sound data in Matlab, we are now saving the FFT directly from the phone app, and using that instead of data processed on the laptop.
Accuracy not great. 88% with 5-fold cross validation, 90% with 80-20 split. Then again there's not a lot of trials.